Cleaned test cases for QR rendering
